requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
lxml==4.6.3
mysql-connector-python==8.0.25
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.3
requests==2.25.1
pymongo==3.12.0